Refactor common component into `components/common`
