+++ src/components/common/mod.rs
@@ -1,0 +1,9 @@
+pub mod breadcrumbs;
+pub mod button;
+pub mod card;
+pub mod data_row;
+pub mod icon_button;
+pub mod list;
+pub mod table_view;
+pub mod toggle_switch;
+pub mod traits;
