+++ src/components/common/traits.rs
